fix: fixed exports and types
