+++ .gitignore
@@ -10,6 +10,7 @@ logs
 *.log
 pnpm-debug.log*
 build/
+.turbo
 
 # Sentry
 .sentryclirc
