+++ examples/next/package.json
@@ -14,7 +14,7 @@
     "@radix-ui/react-popover": "^1.0.7",
     "clsx": "^2.1.0",
     "livepeer": "^3.0.2",
-    "lucide-react": "^0.320.0",
+    "lucide-react": "^0.321.0",
     "next": "14.1.0",
     "react": "^18",
     "react-dom": "^18",
@@ -23,8 +23,8 @@
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
-    "@types/node": "^20.11.13",
-    "@types/react": "^18.2.48",
+    "@types/node": "^20.11.16",
+    "@types/react": "^18.2.51",
     "@types/react-dom": "^18.2.18",
     "autoprefixer": "^10.0.1",
     "postcss": "^8",
