+++ packages/core-react/package.json
@@ -39,7 +39,8 @@
   "scripts": {
     "build": "bunchee",
     "clean": "rimraf .turbo node_modules dist",
-    "dev": "bunchee --watch"
+    "dev": "bunchee --watch",
+    "lint": "tsc --noEmit"
   },
   "peerDependencies": {
     "react": ">=17.0.0"
@@ -52,7 +53,7 @@
   "devDependencies": {
     "@testing-library/react": "^14.2.0",
     "@testing-library/react-hooks": "^8.0.1",
-    "@types/react": "^18.2.48",
+    "@types/react": "^18.2.51",
     "@types/react-dom": "^18.2.18",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
