+++ packages/react/src/broadcast/Controls.tsx
@@ -5,9 +5,9 @@ import { Presence } from "@radix-ui/react-presence";
 import React, { useMemo } from "react";
 
 import { useStore } from "zustand";
-import { MediaScopedProps, useMediaContext } from "../context";
 
 import { useShallow } from "zustand/react/shallow";
+import { MediaScopedProps, useMediaContext } from "../shared/context";
 import * as Radix from "../shared/primitive";
 import { BroadcastScopedProps, useBroadcastContext } from "./context";
