+++ packages/react/src/broadcast/Video.tsx
@@ -5,7 +5,7 @@ import { useStore } from "zustand";
 
 import { addEventListeners } from "@livepeer/core-web/browser";
 import { useComposedRefs } from "@radix-ui/react-compose-refs";
-import { MediaScopedProps, useMediaContext } from "../context";
+import { MediaScopedProps, useMediaContext } from "../shared/context";
 
 import { addBroadcastEventListeners } from "@livepeer/core-web/broadcast";
 import * as Radix from "../shared/primitive";
