+++ packages/react/src/player/Controls.tsx
@@ -8,7 +8,7 @@ import { useStore } from "zustand";
 
 import { composeEventHandlers } from "@radix-ui/primitive";
 import { useShallow } from "zustand/react/shallow";
-import { MediaScopedProps, useMediaContext } from "../context";
+import { MediaScopedProps, useMediaContext } from "../shared/context";
 import * as Radix from "../shared/primitive";
 import { noPropagate } from "../shared/utils";
