+++ packages/react/src/player/MuteTrigger.tsx
@@ -5,7 +5,7 @@ import { composeEventHandlers } from "@radix-ui/primitive";
 import React from "react";
 
 import { useStore } from "zustand";
-import { MediaScopedProps, useMediaContext } from "../context";
+import { MediaScopedProps, useMediaContext } from "../shared/context";
 
 import { useShallow } from "zustand/react/shallow";
 import * as Radix from "../shared/primitive";
