+++ packages/react/src/player/Player.tsx
@@ -12,7 +12,11 @@ import {
 } from "@livepeer/core/media";
 import { createStorage, noopStorage } from "@livepeer/core/storage";
 import { version } from "@livepeer/core/version";
-import { MediaProvider, MediaScopedProps } from "../context";
+import {
+  MediaProvider,
+  MediaScopedProps,
+  useMediaContext,
+} from "../shared/context";
 
 interface PlayerProps
   extends PropsWithChildren<
