+++ packages/react/src/player/Seek.tsx
@@ -7,9 +7,9 @@ import React from "react";
 
 import { useStore } from "zustand";
 import { useShallow } from "zustand/react/shallow";
-import { MediaScopedProps, useMediaContext } from "../context";
 
 import { Presence } from "@radix-ui/react-presence";
+import { MediaScopedProps, useMediaContext } from "../shared/context";
 import * as Radix from "../shared/primitive";
 import { noPropagate } from "../shared/utils";
