+++ packages/react/src/player/VideoQualitySelect.tsx
@@ -5,11 +5,11 @@ import * as SelectPrimitive from "../shared/Select";
 import React from "react";
 
 import { useStore } from "zustand";
-import { MediaScopedProps, useMediaContext } from "../context";
+import { MediaScopedProps, useMediaContext } from "../shared/context";
 
+import { VideoQuality } from "@livepeer/core/media";
 import { composeEventHandlers } from "@radix-ui/primitive";
 import { useShallow } from "zustand/react/shallow";
-import { VideoQuality } from "..";
 import * as Radix from "../shared/primitive";
 
 /**
