+++ packages/react/src/player/Volume.tsx
@@ -3,7 +3,7 @@
 import React, { useEffect, useMemo } from "react";
 
 import { useStore } from "zustand";
-import { MediaScopedProps, useMediaContext } from "../context";
+import { MediaScopedProps, useMediaContext } from "../shared/context";
 
 import { Presence } from "@radix-ui/react-presence";
 import { useShallow } from "zustand/react/shallow";
