+++ packages/react/src/shared/LiveIndicator.tsx
@@ -4,7 +4,7 @@ import React, { useMemo } from "react";
 
 import { Presence } from "@radix-ui/react-presence";
 import { useStore } from "zustand";
-import { MediaScopedProps, useMediaContext } from "../context";
+import { MediaScopedProps, useMediaContext } from "./context";
 import * as Radix from "./primitive";
 
 const LIVE_INDICATOR_NAME = "LiveIndicator";
