+++ packages/react/src/shared/LoadingIndicator.tsx
@@ -4,7 +4,7 @@ import React, { useMemo } from "react";
 
 import { Presence } from "@radix-ui/react-presence";
 import { useStore } from "zustand";
-import { MediaScopedProps, useMediaContext } from "../context";
+import { MediaScopedProps, useMediaContext } from "./context";
 import * as Radix from "./primitive";
 
 const LOADING_INDICATOR_NAME = "LoadingIndicator";
