+++ packages/react/src/shared/Time.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 
 import { useStore } from "zustand";
-import { MediaScopedProps, useMediaContext } from "../context";
+import { MediaScopedProps, useMediaContext } from "./context";
 
 import { useShallow } from "zustand/react/shallow";
 import * as Radix from "./primitive";
